Format env variables as table
This documents which env variables are optional (resolves #61 ), based on my recent experience trying to use this. 

I reformatted some of the inline metadata (allowed values, and default values) as separate columns, only using previously provided content. 

Something that was confusing to me when i was first trying to get this running a few days ago, and remains confusing - what does `default` mean when the variable is required? For example, `ALLOW_DIRECTORY_LIST` claims to a default value of `false`, but i was required to provide it.