Assume the role specified in AWS_ROLE_ARN if AWS_WEB_IDENTITY_TOKEN_FILE is missing
Allow use of AssumeRole to assume the desired identify when web identity is not in use

Also update some signature code to explicitly pass through the desired method instead of assuming it matches the request.
Finally, make it more obvious in signature code that signed components are path rather than a full uri.