feat: Add new configuration option STRIP_LEADING_DIRECTORY_PATH. (#185)
Many thanks to https://github.com/elJosho for this contribution.

This will strip the leading directory from a request if configured.  It's useful when you're trying to expose a bucket at a specific path on an ALB.  For example, if I have an ALB rule that forwards all traffic from "www.mysite.com/mybucket" to the S3 gateway, and I want to request a file in the root of my bucket, "myfile.txt".  Making a request to "www.mysite.com/mybucket/myfile.txt" will fail as the S3 gateway will look for the file inside a folder named "mybucket", rather than at the root of the bucket.  

After this change, setting the new env var to
`STRIP_LEADING_DIRECTORY_PATH = /mybucket`
will send all requests to the root of the bucket.

---------

Co-authored-by: Josh Watt <josh.watt@moodys.com>