feat: Add option to specify a prefix to be prepended to object paths (#190)
Adds an optional configuration environment variable `PREFIX_LEADING_DIRECTORY_PATH`. If defined, this prefix is prepended to all S3 object paths. This is intended as a simple way to allow only some subset of a bucket to be served by the gateway. When used in combination with the existing STRIP_LEADING_DIRECTORY_PATH option, this effectively allows the leading prefix to be replaced, rather than just removed.

Special thanks to @hoffmanr-cshs for this thoughtful contribution.
---------

Co-authored-by: Javier Evans <j.evans@f5.com>