feat: Add new ENV config HEADER_PREFIXES_ALLOWED (#241)
Adds the ability to create a list of header prefixes that are allowed to be returned to the client using the new `HEADER_PREFIXES_ALLOWED` configuration option which is a companion configuration option to `HEADER_PREFIXES_TO_STRIP`.

If specified, this configuration value will take precedence over the default stripped header prefixes (`x-amz-`) as well as and prefixes specified in `HEADER_PREFIXES_TO_STRIP`.

Note that it is generally not advised to return any object store specific headers to the client and the default remains to enforce stripping of these headers.  Use this configuration option with caution.